pyproject.toml
[tool.poetry]
name = "Projeto Advocacia"
version = "0.1.0"
description = ""
authors = ["Ismael Benjamim <ismaelbenjamim15@hotmail.com>"]
license = "CC BY-NC-ND"

[tool.poetry.dependencies]
python = "^3.9"
Django = "^3.2"
python-decouple = "^3.4"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.6"
gunicorn = "^20.1.0"
djangorestframework = "^3.13.1"
django-rest-auth = "^0.9.5"
Pillow = "^9.2.0"



[tool.poetry.dev-dependencies]
pytest-django = "^4.2.0"
flake8 = "^3.9.1"
pytest-cov = "^2.11.1"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "projeto_advocacia.settings"
python_files = [
    "test*.py",
     "*tests.py",
]
minversion = "6.0"
addopts = "-ra -q"
testpaths = [
    "projeto_advocacia/core/",
]


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
